requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary>=2.9.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
supabase==1.0.4
python-multipart==0.0.6
Pillow==10.0.1
gunicorn==21.2.0
Werkzeug==2.3.7
